Create a String class and add function to make string to upper.
Create a String class Add a member
function called upit() that converts the string to all uppercase. and display() to display the string.
You can use the toupper() library function, which takes a single character as an argument and returns a
character that has been converted (if necessary) to uppercase.
Write some code in main() to test upit().